Refactor the rest of the UI
