--- front-end/src/App.css
@@ -1,19 +1,0 @@
-.App {
-	height: 100vh;
-	overflow-y: hidden;
-
-	display: flex;
-	align-items: center;
-
-	background-color: black;
-}
-
-.App-container {
-	/* ipad dimensions */
-	height: 768px;
-	width: 1024px;
-
-	margin: 0 auto;
-
-	background-color: white;
-}
