+++ front-end/src/components/Chat.module.css
@@ -1,5 +1,5 @@
-.responses {
-	width: 40%;
+.chat {
+	width: 60%;
 	height: 100%;
 
 	display: flex;
@@ -21,32 +21,17 @@
 	align-items: center;
 }
 
-.responsesList {
-	background-color: #D9D9D9;
+.messagesList {
+	background-color: white;
+	font-size: 36px;
 
 	width: 100%;
 	flex-grow: 1;
-	padding: 20px;
 	box-sizing: border-box;
+	padding: 20px;
 
 	display: flex;
 	flex-direction: column;
-	justify-content: space-evenly;
-	align-items: center;
-	gap: 20px;
-}
-
-.response {
-	background: #FEFEFE;
-	color: #515151;
-	border: 2px solid #007AFF;
-	font-size: 16px;
-	text-align: left;
-
-	width: 100%;
-	height: 100%;
-	flex-grow: 1;
-	border-radius: 20px;
-	padding: 20px;
-	box-sizing: border-box;
+	justify-content: flex-start;
+	gap: 15px;
 }
